+++ fuel-merkle/src/common/storage_map.rs
@@ -36,6 +36,14 @@ where
             map: Default::default(),
         }
     }
+
+    pub fn is_empty(&self) -> bool {
+        self.map.is_empty()
+    }
+
+    pub fn len(&self) -> usize {
+        self.map.len()
+    }
 }
 
 impl<Type> StorageInspect<Type> for StorageMap<Type>