+++ fuel-tx/src/transaction/consensus_parameters/gas.rs
@@ -322,6 +322,7 @@ pub struct GasCostsValues {
     pub contract_root: DependentCost,
     pub state_root: DependentCost,
     pub new_storage_per_byte: Word, // TODO: this is currently per slot
+    pub vm_initialization: Word,
 }
 
 /// Dependent cost is a cost that depends on the number of units.
@@ -473,9 +474,12 @@ impl GasCostsValues {
             smo: DependentCost::free(),
             srwq: DependentCost::free(),
             swwq: DependentCost::free(),
+
+            // Non-opcode costs
             contract_root: DependentCost::free(),
             state_root: DependentCost::free(),
             new_storage_per_byte: 0,
+            vm_initialization: 0,
         }
     }
 
@@ -588,9 +592,12 @@ impl GasCostsValues {
             smo: DependentCost::unit(),
             srwq: DependentCost::unit(),
             swwq: DependentCost::unit(),
+
+            // Non-opcode costs
             contract_root: DependentCost::unit(),
             state_root: DependentCost::unit(),
             new_storage_per_byte: 1,
+            vm_initialization: 1,
         }
     }
 }