+++ fuel-tx/src/transaction/policies.rs
@@ -1,3 +1,4 @@
+use core::ops::Deref;
 use fuel_types::{
     canonical::{
         Deserialize,
@@ -126,7 +127,7 @@ impl Policies {
 
     /// Sets the `maturity` policy.
     pub fn with_maturity(mut self, maturity: BlockHeight) -> Self {
-        self.set(PolicyType::Maturity, Some(maturity.as_usize() as u64));
+        self.set(PolicyType::Maturity, Some(*maturity.deref() as u64));
         self
     }