+++ fuel-vm/src/checked_transaction.rs
@@ -8,11 +8,11 @@
 #![allow(non_upper_case_globals)]
 
 use fuel_tx::{
-    CheckError,
     Create,
     Mint,
     Script,
     Transaction,
+    ValidityError,
 };
 use fuel_types::{
     BlockHeight,
@@ -170,6 +170,16 @@ impl<Tx: IntoChecked> Borrow<Tx> for Checked<Tx> {
     fn borrow(&self) -> &Tx {
         self.transaction()
     }
+}
+
+/// The error can occur when transforming transactions into the `Checked` type.
+#[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub enum CheckError {
+    /// The transaction doesn't pass validity rules.
+    Validity(ValidityError),
+    /// The predicate verification failed.
+    PredicateVerificationFailed(PredicateVerificationFailed),
 }
 
 /// Performs checks for a transaction
@@ -605,6 +615,18 @@ impl IntoChecked for Transaction {
     }
 }
 
+impl From<ValidityError> for CheckError {
+    fn from(value: ValidityError) -> Self {
+        CheckError::Validity(value)
+    }
+}
+
+impl From<PredicateVerificationFailed> for CheckError {
+    fn from(value: PredicateVerificationFailed) -> Self {
+        CheckError::PredicateVerificationFailed(value)
+    }
+}
+
 #[cfg(feature = "random")]
 #[cfg(test)]
 mod tests {
@@ -619,9 +641,9 @@ mod tests {
             WitnessLimit,
             Witnesses,
         },
-        CheckError,
         Script,
         TransactionBuilder,
+        ValidityError,
     };
     use fuel_types::canonical::Serialize;
     use quickcheck::TestResult;
@@ -731,10 +753,10 @@ mod tests {
         // verify available balance was decreased by max fee
         assert!(matches!(
             err,
-            CheckError::InsufficientFeeAmount {
+            CheckError::Validity(ValidityError::InsufficientFeeAmount {
                 expected: _,
                 provided: 0
-            }
+            })
         ));
     }
 
@@ -769,10 +791,10 @@ mod tests {
         // verify available balance was decreased by max fee
         assert!(matches!(
             err,
-            CheckError::InsufficientFeeAmount {
+            CheckError::Validity(ValidityError::InsufficientFeeAmount {
                 expected: _,
                 provided: 0
-            }
+            })
         ));
     }
 
@@ -987,6 +1009,7 @@ mod tests {
 
         let min_fee = fee.min_fee();
         let expected_min_fee = (tx.metered_bytes_size() as u64 * fee_params.gas_per_byte
+            + gas_costs.vm_initialization
             + 3 * gas_costs.ecr1
             + gas_costs.s256.resolve(tx.size() as u64))
             * gas_price;
@@ -1038,6 +1061,7 @@ mod tests {
         // be recovered once. Therefore, we charge only once for the address
         // recovery of the signed inputs.
         let expected_min_fee = (tx.metered_bytes_size() as u64 * fee_params.gas_per_byte
+            + gas_costs.vm_initialization
             + gas_costs.ecr1
             + gas_costs.s256.resolve(tx.size() as u64))
             * gas_price;
@@ -1102,9 +1126,11 @@ mod tests {
 
         let min_fee = fee.min_fee();
         let expected_min_fee = (tx.metered_bytes_size() as u64 * fee_params.gas_per_byte
+            + gas_costs.vm_initialization
             + gas_costs.contract_root.resolve(predicate_1.len() as u64)
             + gas_costs.contract_root.resolve(predicate_2.len() as u64)
             + gas_costs.contract_root.resolve(predicate_3.len() as u64)
+            + 3 * gas_costs.vm_initialization
             + 50
             + 100
             + 200
@@ -1186,9 +1212,11 @@ mod tests {
         let min_fee = fee.min_fee();
         let expected_min_fee = (tx.metered_bytes_size() as u64 * fee_params.gas_per_byte
             + 3 * gas_costs.ecr1
+            + gas_costs.vm_initialization
             + gas_costs.contract_root.resolve(predicate_1.len() as u64)
             + gas_costs.contract_root.resolve(predicate_2.len() as u64)
             + gas_costs.contract_root.resolve(predicate_3.len() as u64)
+            + 3 * gas_costs.vm_initialization
             + 50
             + 100
             + 200
@@ -1226,6 +1254,7 @@ mod tests {
         let expected_min_fee = (tx.metered_bytes_size() as u64 * fee_params.gas_per_byte
             + gas_costs.state_root.resolve(storage_slots_len as Word)
             + gas_costs.contract_root.resolve(bytecode_len as Word)
+            + gas_costs.vm_initialization
             + gas_costs.s256.resolve(100)
             + gas_costs.s256.resolve(tx.size() as u64))
             * gas_price;
@@ -1258,6 +1287,7 @@ mod tests {
         let expected_min_fee = (tx.metered_bytes_size() as u64 * fee_params.gas_per_byte
             + gas_costs.state_root.resolve(0)
             + gas_costs.contract_root.resolve(0)
+            + gas_costs.vm_initialization
             + gas_costs.s256.resolve(100)
             + gas_costs.s256.resolve(tx.size() as u64))
             * gas_price;
@@ -1311,7 +1341,10 @@ mod tests {
             .expect_err("Expected invalid transaction");
 
         // assert that tx without base input assets fails
-        assert!(matches!(checked, CheckError::InsufficientFeeAmount { .. }));
+        assert!(matches!(
+            checked,
+            CheckError::Validity(ValidityError::InsufficientFeeAmount { .. })
+        ));
     }
 
     #[test]
@@ -1332,7 +1365,10 @@ mod tests {
             .expect_err("overflow expected");
 
         let provided = match err {
-            CheckError::InsufficientFeeAmount { provided, .. } => provided,
+            CheckError::Validity(ValidityError::InsufficientFeeAmount {
+                provided,
+                ..
+            }) => provided,
             _ => panic!("expected insufficient fee amount; found {err:?}"),
         };
 
@@ -1358,7 +1394,10 @@ mod tests {
             .expect_err("overflow expected");
 
         let provided = match err {
-            CheckError::InsufficientFeeAmount { provided, .. } => provided,
+            CheckError::Validity(ValidityError::InsufficientFeeAmount {
+                provided,
+                ..
+            }) => provided,
             _ => panic!("expected insufficient fee amount; found {err:?}"),
         };
 
@@ -1380,7 +1419,7 @@ mod tests {
             .into_checked(Default::default(), &consensus_params)
             .expect_err("overflow expected");
 
-        assert_eq!(err, CheckError::ArithmeticOverflow);
+        assert_eq!(err, CheckError::Validity(ValidityError::BalanceOverflow));
     }
 
     #[test]
@@ -1398,7 +1437,7 @@ mod tests {
             .into_checked(Default::default(), &consensus_params)
             .expect_err("overflow expected");
 
-        assert_eq!(err, CheckError::ArithmeticOverflow);
+        assert_eq!(err, CheckError::Validity(ValidityError::BalanceOverflow));
     }
 
     #[test]
@@ -1438,11 +1477,11 @@ mod tests {
             .expect_err("Expected valid transaction");
 
         assert_eq!(
-            CheckError::InsufficientInputAmount {
+            CheckError::Validity(ValidityError::InsufficientInputAmount {
                 asset: any_asset,
                 expected: input_amount + 1,
                 provided: input_amount
-            },
+            }),
             checked
         );
     }
@@ -1518,7 +1557,7 @@ mod tests {
         gas_costs: &GasCosts,
         fee_params: &FeeParameters,
         base_asset_id: &AssetId,
-    ) -> Result<bool, CheckError> {
+    ) -> Result<bool, ValidityError> {
         fn gas_to_fee(gas: u64, price: u64, factor: u64) -> u128 {
             let prices_gas = gas as u128 * price as u128;
             let fee = prices_gas / factor as u128;
@@ -1536,7 +1575,8 @@ mod tests {
         let gas_used_by_metadata = tx.gas_used_by_metadata(gas_costs);
         let min_gas = gas_used_by_bytes
             .saturating_add(gas_used_by_inputs)
-            .saturating_add(gas_used_by_metadata);
+            .saturating_add(gas_used_by_metadata)
+            .saturating_add(gas_costs.vm_initialization);
 
         // use different division mechanism than impl
         let witness_limit_allowance = tx
@@ -1548,7 +1588,7 @@ mod tests {
             .saturating_add(witness_limit_allowance);
         let max_fee: u64 = gas_to_fee(max_gas, tx.price(), fee_params.gas_price_factor)
             .try_into()
-            .map_err(|_| CheckError::ArithmeticOverflow)?;
+            .map_err(|_| ValidityError::BalanceOverflow)?;
 
         let result = max_fee == available_balances.fee.max_fee();
         Ok(result)
@@ -1559,7 +1599,7 @@ mod tests {
         gas_costs: &GasCosts,
         fee_params: &FeeParameters,
         base_asset_id: &AssetId,
-    ) -> Result<bool, CheckError>
+    ) -> Result<bool, ValidityError>
     where
         Tx: Chargeable + field::Inputs + field::Outputs,
     {
@@ -1574,7 +1614,8 @@ mod tests {
         let gas_used_by_metadata = tx.gas_used_by_metadata(gas_costs);
         let gas = gas_used_by_bytes
             .saturating_add(gas_used_by_inputs)
-            .saturating_add(gas_used_by_metadata);
+            .saturating_add(gas_used_by_metadata)
+            .saturating_add(gas_costs.vm_initialization);
         let total = gas as u128 * tx.price() as u128;
         // use different division mechanism than impl
         let fee = total / fee_params.gas_price_factor as u128;
@@ -1583,7 +1624,7 @@ mod tests {
         let rounded_fee = fee.saturating_add(fee_remainder);
         let min_fee: u64 = rounded_fee
             .try_into()
-            .map_err(|_| CheckError::ArithmeticOverflow)?;
+            .map_err(|_| ValidityError::BalanceOverflow)?;
 
         Ok(min_fee == available_balances.fee.min_fee())
     }