+++ fuel-vm/src/interpreter/metadata.rs
@@ -8,6 +8,7 @@ use crate::{
     constraints::reg_key::*,
     consts::*,
     context::Context,
+    convert,
     error::SimpleResult,
 };
 
@@ -147,7 +148,7 @@ impl<Tx> GTFInput<'_, Tx> {
     where
         Tx: ExecutableTransaction,
     {
-        let b = usize::try_from(b).map_err(|_| PanicReason::ArithmeticOverflow)?;
+        let b = convert::to_usize(b).ok_or(PanicReason::InvalidMetadataIdentifier)?;
         let args = GTFArgs::try_from(imm)?;
         let tx = self.tx;
         let ofs = self.tx_offset;