+++ fuel-vm/src/tests/predicate.rs
@@ -21,6 +21,7 @@ use crate::{
 };
 
 use crate::checked_transaction::{
+    CheckError,
     CheckPredicateParams,
     CheckPredicates,
     EstimatePredicates,
@@ -226,8 +227,9 @@ async fn execute_gas_metered_predicates(
         );
     }
 
-    for predicate in predicates {
+    for predicate in predicates.iter() {
         let predicate: Vec<u8> = predicate
+            .clone()
             .into_iter()
             .flat_map(|op| u32::from(op).to_be_bytes())
             .collect();
@@ -240,7 +242,7 @@ async fn execute_gas_metered_predicates(
             AssetId::default(),
             rng.gen(),
             Default::default(),
-            GAS_LIMIT,
+            0,
             predicate,
             vec![],
         );
@@ -251,7 +253,6 @@ async fn execute_gas_metered_predicates(
     let mut transaction = builder.finalize();
 
     let params = CheckPredicateParams {
-        max_gas_per_tx: GAS_LIMIT,
         max_gas_per_predicate: GAS_LIMIT,
         ..Default::default()
     };
@@ -552,18 +553,18 @@ async fn gas_used_by_predicates_more_than_limit() {
             .check_predicates_async::<TokioWithRayon>(&params)
             .await;
 
-        assert_eq!(
+        assert!(matches!(
             tx_with_predicate.unwrap_err(),
-            CheckError::PredicateVerificationFailed
-        );
+            CheckError::PredicateVerificationFailed(_)
+        ));
     }
 
     let tx_with_predicate = builder
         .finalize_checked_basic(Default::default())
         .check_predicates(&params);
 
-    assert_eq!(
+    assert!(matches!(
         tx_with_predicate.unwrap_err(),
-        CheckError::PredicateVerificationFailed
-    );
+        CheckError::PredicateVerificationFailed(_)
+    ));
 }