+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 #### Breaking
 
+- [#629](https://github.com/FuelLabs/fuel-vm/pull/629): Charge the user for VM initialization.
+- [#628](https://github.com/FuelLabs/fuel-vm/pull/628): Renamed `transaction::CheckError` to `transaction::ValidityError`.
+  Created a new `checked_transaction::CheckError` that combines `ValidityError`
+  and `PredicateVerificationFailed` errors into one. It allows the return of the
+  `PredicateVerificationFailed` to the end user instead of losing the reason why predicate verification failed.
+- [#625](https://github.com/FuelLabs/fuel-vm/pull/625): Use `ArithmeticError` only for arithmetic operations, and introduce new errors like `BalanceOverflow` for others. Whenever an error is internally caused by a type conversion to `usize`, so that an overflowing value wouldn't map to a valid index anyway, return the missing item error instead.
 - [#623](https://github.com/FuelLabs/fuel-vm/pull/623):
   Added support for transaction policies. The `Script` and `Create`
   transactions received a new field, `policies`. Policies allow the addition
@@ -39,6 +45,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     - If the `max_fee > policies.max_fee`, then transaction will be rejected.
     - If the `witnessses_size > policies.witness_limit`, then transaction will be rejected.
     - GTF opcode changed its hardcoded constants for fields. It should be updated according to the values from the specification on the Sway side.
+- [#633](https://github.com/FuelLabs/fuel-vm/pull/633): Limit receipt count to `u16::MAX`.
+
+### Fixed
+
+- [#627](https://github.com/FuelLabs/fuel-vm/pull/627): Added removal of obsolete SMT nodes along the path during `update` and `delete` operations.
 
 ## [Version 0.41.0]